requirements.txt
Flask==2.3.3
tinydb==4.8.0
tinydb-serialization==2.2.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
scrapy==2.11.0
pandas==2.1.1
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
gunicorn==21.2.0
